requirements.txt
termcolor >= 1.1.0
selenium >= 4.1.3
requests >= 2.27.1
playsound == 1.2.2
win10toast >= 0.9
websocket-client >= 1.3.3
cloudscraper >= 1.2.60
Bloxflip.py >= 2.2.21
